requirements.txt
# VigilHome AI Core Dependencies

# Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.200
onnxruntime>=1.16.0

# Scene Understanding (VLM)
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
accelerate>=0.24.0

# Vector Database & Search
chromadb>=0.4.15
sentence-transformers>=2.2.2

# Time-series & Data
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Anomaly Detection
pyod>=1.1.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Image Processing
Pillow>=10.0.0
imageio>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
schedule>=1.2.0
tqdm>=4.66.0

# Monitoring
psutil>=5.9.0
